Make the menu icon animate and insert the SVG file into the HTML
I made the icon animate with SMIL but also needed to insert the SVG data
into the HTML file.

I tried to use the "document.getElementById" to traget the ID of the svg
elements but it didn't work, because the SVG is in a separate file. This
makes it a problem. I don't have a solution execpt doing some weried
code form a 3-party that I don't understand yet.

WHY
I want it to work right now and then worry about it later.

Signed-off-by: Mateusz Rybinski <rybinmateusz@gmail.com>