Fix the scroll when the menu is longer then the window
Make it possible to scroll when the menu is longer then window.
I needed to use a custom "fixed" class. The foundation provide,
is not the best and I was very lucky to find this solution.
I'm super happy and this solution works really well.

Res: http://stackoverflow.com/questions/16094785/have-a-fixed-position-div-that-needs-to-scroll-if-content-overflows

Signed-of-by: Mateusz Rybinski <rybinmateusz@gmail.com>